pyproject.toml
[tool.poetry]
name = "django-null"
version = "0.1.0"
description = ""
authors = ["usoma-ak <alihonovusoma@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
django-jsonform = "^2.23.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
